% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tlm-package.R
\docType{package}
\name{tlm-package}
\alias{tlm-package}
\title{tlm: Effects under Linear, Logistic and Poisson Regression Models with Transformed Variables}
\description{
Computation of effects under linear, logistic and Poisson regression models with transformed variables. Logarithm and power transformations are allowed. Effects can be displayed both numerically and graphically in both the original and the transformed space of the variables. The methods are described in Barrera-Gomez and Basagana (2015) \doi{10.1097/EDE.0000000000000247}.
}
\author{
\strong{Maintainer}: Jose Barrera-Gomez \email{jose.barrera@isglobal.org} (\href{https://orcid.org/0000-0002-2688-6036}{ORCID})

Authors:
\itemize{
  \item Xavier Basagana \email{xavier.basagana@isglobal.org} (\href{https://orcid.org/0000-0002-8457-1489}{ORCID})
}

}
\keyword{internal}
